flask_app/requirements.txt
flask
torch
torchvision
transformers
qdrant-client
pillow
